Moving GitHub-related services to their own folder/namespace
